created modules for code and added meaningfull names
